Updated after feedback of Martin
