--- resources/js/mui_theme.js
@@ -1,15 +1,0 @@
-
-const theme = MaterialUI.createMuiTheme({ 
-  typography: {
-    useNextVariants: true,
-  },
-  palette: {
-    primary: {
-      light: '#63ccff',
-      main: '#009be5',
-      dark: '#006db3',
-      contrastText: '#fff',
-    },
-  },
- }); 
-
